keen: adding addons, fixes #85 and fixing error responses.
Went with Keen's default names for the addons so that they hopefully
match up with the analytics.js versions. We will still have to
add settings in integration-metadata for it as well, since the
API will reject the request.

Also looks like Keen sends back res.error when problems happen,
which is how I discovered that they reject the request in the
first place.